requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
asttokens==2.4.1
bcrypt==4.2.1
cachetools==5.5.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
colorama==0.4.6
decorator==5.1.1
distro==1.9.0
Django==5.1.4
django-cors-headers==4.6.0
django-dirtyfields==1.9.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
et_xmlfile==2.0.0
executing==2.1.0
google-ai-generativelanguage==0.6.6
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-generativeai==0.7.2
googleapis-common-protos==1.66.0
grpcio==1.69.0
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
ipython==8.29.0
jedi==0.19.2
jiter==0.8.2
matplotlib-inline==0.1.7
mysqlclient==2.2.6
numpy==2.1.3
oauthlib==3.2.2
openai==1.59.4
openpyxl==3.1.5
pandas==2.2.3
parso==0.8.4
pillow==11.0.0
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==4.25.5
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.18.0
PyJWT==2.10.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
reportlab==4.2.5
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.2
stack-data==0.6.3
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
wcwidth==0.2.13

